Revert "fix: refactor skypilot launch system (#2198)"
This reverts commit 9f27d4188a1ba2566d01e452923438d9879cccc3.
