+++ .cspell.json
@@ -227,7 +227,6 @@
     "sessionfinish",
     "setsid",
     "setuptools",
-    "sigval",
     "skylaunch",
     "skypilot",
     "softmax",
