+++ agent/src/metta/agent/metta_agent.py
@@ -15,8 +15,6 @@ from metta.rl.system_config import SystemConfig
 
 logger = logging.getLogger("metta_agent")
 
-IS_MASTER = not torch.distributed.is_initialized() or torch.distributed.get_rank() == 0
-
 
 class DistributedMettaAgent(DistributedDataParallel):
     """
@@ -28,8 +26,7 @@ class DistributedMettaAgent(DistributedDataParallel):
     module: "MettaAgent"
 
     def __init__(self, agent: "MettaAgent", device: torch.device):
-        if IS_MASTER:
-            logger.info("Converting BatchNorm layers to SyncBatchNorm for distributed training...")
+        logger.info("Converting BatchNorm layers to SyncBatchNorm for distributed training...")
 
         # Check if the agent might have circular references that would cause recursion
         # This can happen with legacy checkpoints wrapped in LegacyMettaAgentAdapter
@@ -175,8 +172,7 @@ class MettaAgent(nn.Module):
         # Store original feature mapping on first initialization
         if not hasattr(self, "original_feature_mapping"):
             self.original_feature_mapping = {name: props["id"] for name, props in features.items()}
-            if IS_MASTER:
-                logger.info(f"Stored original feature mapping with {len(self.original_feature_mapping)} features")
+            logger.info(f"Stored original feature mapping with {len(self.original_feature_mapping)} features")
         else:
             # Create remapping for subsequent initializations
             self._create_feature_remapping(features)
@@ -283,9 +279,7 @@ class MettaAgent(nn.Module):
             device=device,
             dtype=torch.int32,
         )
-
-        if IS_MASTER:
-            logger.info(f"Actions initialized: {self.active_actions}")
+        logger.info(f"Actions initialized: {self.active_actions}")
 
         # Pass tensors to policy if needed
         if self.policy is not None:
