+++ devops/skypilot/config/configure_environment.sh
@@ -25,38 +25,37 @@ mkdir -p "$JOB_METADATA_DIR"
 RESTART_COUNT_FILE="$JOB_METADATA_DIR/restart_count"
 ACCUMULATED_RUNTIME_FILE="$JOB_METADATA_DIR/accumulated_runtime"
 CLUSTER_STOP_FILE="$JOB_METADATA_DIR/cluster_stop"
-TERMINATION_REASON_FILE="$JOB_METADATA_DIR/termination_reason"
-HEARTBEAT_FILE="${HEARTBEAT_FILE:-$JOB_METADATA_DIR/heartbeat_file}"
 
 # Initialize or update restart tracking
 if [ -f "$RESTART_COUNT_FILE" ]; then
-  RESTART_COUNT=$(cat "$RESTART_COUNT_FILE")
-  RESTART_COUNT=$((RESTART_COUNT + 1))
+    RESTART_COUNT=$(cat "$RESTART_COUNT_FILE")
+    RESTART_COUNT=$((RESTART_COUNT + 1))
 else
-  RESTART_COUNT=0
+    RESTART_COUNT=0
 fi
 
 if [[ "$IS_MASTER" == "true" ]]; then
-  echo "$RESTART_COUNT" > "$RESTART_COUNT_FILE"
-  # Clear any stale cluster stop flag at the beginning of a fresh attempt
-  : > "$CLUSTER_STOP_FILE" 2> /dev/null || true
+    echo "$RESTART_COUNT" > "$RESTART_COUNT_FILE"
+    # Clear any stale cluster stop flag at the beginning of a fresh attempt
+    : > "$CLUSTER_STOP_FILE" 2>/dev/null || true
 else
-  echo "[INFO] Skipping RESTART_COUNT_FILE and CLUSTER_STOP_FILE updates on non-master node"
+    echo "[INFO] Skipping signal file updates on non-master node"
 fi
 
 # Read accumulated runtime
 if [ -f "$ACCUMULATED_RUNTIME_FILE" ]; then
-  ACCUMULATED_RUNTIME=$(cat "$ACCUMULATED_RUNTIME_FILE")
+    ACCUMULATED_RUNTIME=$(cat "$ACCUMULATED_RUNTIME_FILE")
 else
-  ACCUMULATED_RUNTIME=0
+    ACCUMULATED_RUNTIME=0
 fi
 
-echo "============= RESTART INFO ============="
+
+echo "[RESTART INFO] ========================"
 echo "  METTA_RUN_ID: ${METTA_RUN_ID}"
 echo "  RESTART_COUNT: ${RESTART_COUNT}"
 echo "  ACCUMULATED_RUNTIME: ${ACCUMULATED_RUNTIME}s ($((ACCUMULATED_RUNTIME / 60))m)"
 echo "  METADATA_DIR: ${JOB_METADATA_DIR}"
-echo "========================================"
+echo "=================================="
 
 # Write all environment variables using heredoc
 cat >> "$METTA_ENV_FILE" << EOF
@@ -77,24 +76,24 @@ export NODE_INDEX="\${SKYPILOT_NODE_RANK}"
 # Job metadata exports
 export RESTART_COUNT="${RESTART_COUNT}"
 export ACCUMULATED_RUNTIME="${ACCUMULATED_RUNTIME}"
-
-# File path exports for monitors
-export JOB_METADATA_DIR="${JOB_METADATA_DIR}"
 export ACCUMULATED_RUNTIME_FILE="${ACCUMULATED_RUNTIME_FILE}"
 export CLUSTER_STOP_FILE="${CLUSTER_STOP_FILE}"
-export HEARTBEAT_FILE="${HEARTBEAT_FILE}"
-export TERMINATION_REASON_FILE="${TERMINATION_REASON_FILE}"
 
 # NCCL Configuration
 export NCCL_PORT_RANGE="\${NCCL_PORT_RANGE:-43000-43063}"
+export NCCL_SOCKET_IFNAME="\${NCCL_SOCKET_IFNAME:-enp39s0}"
 export NCCL_SOCKET_FAMILY="\${NCCL_SOCKET_FAMILY:-AF_INET}"
+export NCCL_MIN_NCHANNELS="\${NCCL_MIN_NCHANNELS:-4}"
+export NCCL_MAX_NCHANNELS="\${NCCL_MAX_NCHANNELS:-8}"
+export NCCL_SOCKET_NTHREADS="\${NCCL_SOCKET_NTHREADS:-2}"
+export NCCL_NSOCKS_PERTHREAD="\${NCCL_NSOCKS_PERTHREAD:-4}"
 
 # Debug
 export TORCH_NCCL_ASYNC_ERROR_HANDLING="\${TORCH_NCCL_ASYNC_ERROR_HANDLING:-1}"
-export NCCL_DEBUG=WARN
-export NCCL_DEBUG_SUBSYS=""
+export NCCL_DEBUG=VERSION
+export NCCL_DEBUG_SUBSYS=INIT,IPC
+export CUDA_LAUNCH_BLOCKING=1
 
-# NCCL Mode
 export NCCL_P2P_DISABLE="\${NCCL_P2P_DISABLE:-0}"
 export NCCL_SHM_DISABLE="\${NCCL_SHM_DISABLE:-0}"
 export NCCL_IB_DISABLE="\${NCCL_IB_DISABLE:-1}"
@@ -103,25 +102,25 @@ EOF
 
 # Create job secrets (idempotent - overwrites if exists)
 if [ -z "$WANDB_PASSWORD" ]; then
-  echo "ERROR: WANDB_PASSWORD environment variable is required but not set"
-  echo "Please ensure WANDB_PASSWORD is set in your Skypilot environment variables"
-  exit 1
+    echo "ERROR: WANDB_PASSWORD environment variable is required but not set"
+    echo "Please ensure WANDB_PASSWORD is set in your Skypilot environment variables"
+    exit 1
 fi
 
 echo "Creating/updating job secrets..."
 
 # Build command - wandb-password is always included
-CMD="uv run ./devops/skypilot/config/lifecycle/create_job_secrets.py --wandb-password \"$WANDB_PASSWORD\""
+CMD="uv run ./devops/skypilot/create_job_secrets.py --wandb-password \"$WANDB_PASSWORD\""
 
 # Add observatory-token only if it's set
 if [ -n "$OBSERVATORY_TOKEN" ]; then
-  CMD="$CMD --observatory-token \"$OBSERVATORY_TOKEN\""
+    CMD="$CMD --observatory-token \"$OBSERVATORY_TOKEN\""
 fi
 
 # Execute the command
 eval $CMD || {
-  echo "ERROR: Failed to create job secrets"
-  exit 1
+    echo "ERROR: Failed to create job secrets"
+    exit 1
 }
 
 echo "Runtime environment configuration completed"
