+++ devops/skypilot/config/monitors/cluster_stop_monitor.sh
@@ -5,23 +5,20 @@ set -euo pipefail
 : "${CMD_PID:?Missing CMD_PID}"
 : "${WRAPPER_PID:?Missing WRAPPER_PID}"
 : "${CLUSTER_STOP_FILE:?Missing CLUSTER_STOP_FILE}"
+: "${CLUSTER_STOP_CHECK_INTERVAL:?Missing CLUSTER_STOP_CHECK_INTERVAL}"
 : "${TERMINATION_REASON_FILE:?Missing TERMINATION_REASON_FILE}"
-
-CLUSTER_STOP_CHECK_INTERVAL=${CLUSTER_STOP_CHECK_INTERVAL:-15}
 
 echo "[INFO] Cluster-stop monitor started; checking every ${CLUSTER_STOP_CHECK_INTERVAL}s"
 
-while true; do
-  sleep "$CLUSTER_STOP_CHECK_INTERVAL"
-
+while kill -0 "$CMD_PID" 2>/dev/null; do
   if [ -s "$CLUSTER_STOP_FILE" ]; then
-    reason="$(cat "$CLUSTER_STOP_FILE" 2> /dev/null || true)"
+    reason="$(cat "$CLUSTER_STOP_FILE" 2>/dev/null || true)"
     echo "[INFO] Cluster stop flag detected (${reason:-no-reason}); requesting shutdown"
     echo "${reason:-cluster_stop}" > "$TERMINATION_REASON_FILE"
-    kill -TERM "${WRAPPER_PID}" 2> /dev/null || true
+    kill -TERM "${WRAPPER_PID}" 2>/dev/null || true
     break
   fi
-
+  sleep "$CLUSTER_STOP_CHECK_INTERVAL"
 done
 
 echo "[INFO] Cluster-stop monitor exiting"
