+++ metta/rl/trainer.py
@@ -298,8 +298,7 @@ def train(
 
     # Wrap in DDP if distributed
     if torch.distributed.is_initialized():
-        if torch.distributed.get_rank() == 0:
-            logger.info("Initializing DistributedDataParallel")
+        logger.info(f"Initializing DistributedDataParallel on device {device}")
         torch.distributed.barrier()
         policy = wrap_agent_distributed(policy, device)
         torch.distributed.barrier()
